+++ src/main/webapp/app/src/App.jsx
@@ -2,12 +2,12 @@ import * as React from 'react'
 import { useEffect, useState } from 'react'
 import CssBaseline from '@mui/material/CssBaseline'
 import { BrowserRouter, Route, Routes } from 'react-router-dom'
-import Footer from './components/Footer'
-import NavBar from './components/NavBar'
+import Footer from './components/core/Footer'
+import NavBar from './components/core/NavBar'
 import Container from '@mui/material/Container'
 import AppContext, { defaultAppContext } from './core/AppContext'
 import { ErrorBoundary } from 'react-error-boundary'
-import ErrorFallback from './components/ErrorFallback'
+import ErrorFallback from './components/core/ErrorFallback'
 
 import './App.css'
