+++ src/main/webapp/app/src/components/dashboard/PendingResponseGrid.jsx
@@ -7,13 +7,13 @@ import {
   Inject,
   Toolbar,
 } from '@syncfusion/ej2-react-grids'
-import RestAdaptor from '../core/RestAdaptor'
+import RestAdaptor from '../../core/RestAdaptor'
 
-import { API, ROUTES, TEXT, TOOLBAR } from '../core/const'
+import { API, ROUTES, TEXT, TOOLBAR } from '../../core/const'
 import { useContext } from 'react'
-import AppContext from '../core/AppContext'
+import AppContext from '../../core/AppContext'
 import { Link } from 'react-router-dom'
-import { fromTemplate } from '../core/util'
+import { fromTemplate } from '../../core/util'
 import MaterialLink from '@mui/material/Link'
 
 const PendingResponseGrid = () => {
