+++ src/main/webapp/app/src/components/schedule/InvitationGrid.jsx
@@ -7,11 +7,11 @@ import {
   Inject,
   Toolbar,
 } from '@syncfusion/ej2-react-grids'
-import RestAdaptor from '../core/RestAdaptor'
+import RestAdaptor from '../../core/RestAdaptor'
 
-import { API, ERROR_MESSAGES, TOOLBAR } from '../core/const'
+import { API, ERROR_MESSAGES, TOOLBAR } from '../../core/const'
 import Typography from '@mui/material/Typography'
-import { fromTemplate } from '../core/util'
+import { fromTemplate } from '../../core/util'
 
 const InvitationGrid = ({ rosterPlan, readonly }) => {
   const template = rosterPlan._links.rosterPlanUserInfos.href
