+++ src/main/webapp/app/src/components/schedule/owner/ScheduleOwner.jsx
@@ -10,12 +10,13 @@ import {
 } from '@syncfusion/ej2-react-schedule'
 import * as React from 'react'
 import { DataManager } from '@syncfusion/ej2-data'
-import RestAdaptor from '../core/RestAdaptor'
-import { API } from '../core/const'
-import ScheduleEditEditorTemplate from './ScheduleEditEditorTemplate'
-import { fromTemplate } from '../core/util'
+import RestAdaptor from '../../../core/RestAdaptor'
+import { API } from '../../../core/const'
+import ScheduleOwnerEditorTemplate from './ScheduleOwnerEditorTemplate'
+import { fromTemplate } from '../../../core/util'
+import ScheduleEventFooter from '../ScheduleEventFooter'
 
-const ScheduleEdit = ({ rosterPlan }) => {
+const ScheduleOwner = ({ rosterPlan }) => {
   const template = rosterPlan?._links?.timeSlots?.href
   if (template === null) {
     throw new Error('Meeting not found.')
@@ -46,11 +47,17 @@ const ScheduleEdit = ({ rosterPlan }) => {
     }
   }
 
+  const onEventRendered = ({ data: { timeSlotUserInfos }, element }) => {
+    const [appointment] = element.children
+    appointment.append(ScheduleEventFooter(timeSlotUserInfos))
+  }
+
   return (
     <ScheduleComponent
-      editorTemplate={ScheduleEditEditorTemplate}
+      editorTemplate={ScheduleOwnerEditorTemplate}
       eventSettings={eventSettings}
       height='80vh'
+      eventRendered={onEventRendered}
     >
       <Inject
         services={[Day, Week, Month, Agenda, Resize, DragAndDrop]}
@@ -59,4 +66,4 @@ const ScheduleEdit = ({ rosterPlan }) => {
   )
 }
 
-export default ScheduleEdit
+export default ScheduleOwner
