+++ src/main/webapp/app/src/components/schedule/owner/SolutionGrid.jsx
@@ -7,14 +7,14 @@ import {
   Inject,
   Toolbar,
 } from '@syncfusion/ej2-react-grids'
-import RestAdaptor from '../core/RestAdaptor'
+import RestAdaptor from '../../../core/RestAdaptor'
 
-import { API, ROUTES } from '../core/const'
+import { API, ROUTES } from '../../../core/const'
 import { Link } from 'react-router-dom'
 import Button from '@mui/material/Button'
 import * as React from 'react'
 import { useRef } from 'react'
-import ajax from '../core/ajax'
+import ajax from '../../../core/ajax'
 
 const SolutionGrid = ({ rosterPlan }) => {
   const ref = useRef(null)
