+++ src/main/webapp/app/src/components/schedule/owner/quick_menu/QuickMenuEventSettings.jsx
