+++ src/main/webapp/app/src/pages/Profile.jsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import { useContext } from 'react'
-import CenterWrapper from '../components/CenterWrapper'
+import CenterWrapper from '../components/core/CenterWrapper'
 import AppContext from '../core/AppContext'
 import FormControl from '@mui/material/FormControl'
 import { FormLabel } from '@mui/material'
